finalize changes for cell health figure
